engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "reference-to-rig"
version = "0.1.0"
description = "Local guitar tone matching and rig recipe generator"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "pydub>=0.25.1",
    "pyloudnorm>=0.1.1",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "faiss-cpu>=1.7.4",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "tenacity>=8.2.3",
    "cma>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"


